+++ Toggly.FeatureManagement.ReactNative/libs/react-native/src/hooks/index.ts
@@ -7,3 +7,5 @@ export type {
 
 export { useToggly } from './useToggly';
 export type { UseTogglyResult } from './useToggly';
+
+export { useVariant } from './useVariant';